\name{RcmdrPlugin.depthTools-package}
\alias{RcmdrPlugin.depthTools-package}
\alias{RcmdrPlugin.depthTools}
\docType{package}
\title{
Graphical User Interface for depthTools 
}
\description{
This package provides an Rcmdr "plug-in" based on the depthTools package. It provides a GUI for the computation and graphical visualization of the Modified Band Depth (MBD) of points in a data set with respect to the same or a different set. It also allows to compute the scale curve of a set, to perform the rank test to decide whether two samples come from the same population, and to classify new data points according to the DS and the TAD classification methods. In addition, the user can obtain a plot which enhances a fixed percentage of the most central curves through the centralPlot, and compute a trimmed mean, based on MBD. 
}
\details{
\tabular{ll}{
Package: RcmdrPlugin.dephTools
Type: Package
Version: 1.2
Date: 2013-02-14
License: GPL (>= 2)
}
}

\author{
Aurora Torrente <etorrent@est-econ.uc3m.es> and Sara Lopez-Pintado <sl2929@columbia.edu>

Maintainer: Aurora Torrente <etorrent@est-econ.uc3m.es>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

